+++ CHANGELOG.md
@@ -36,6 +36,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   directory and from `routes/console.php` if it exists without needing to define
   a `commands()` method in your application's console kernel.
 
+### Fixed
+
+- Ensure that multiple HTTP request calls properly load the
+  header/footer/sidebar template. Previously only the first request would load
+  the header/footer/sidebar templates and the response returned was only the
+  template content.
+
 ## v1.6.0
 
 ### Added